backend/requirements.txt
flask
flask_socketio
python-dotenv
eventlet
termcolor